Restructuring jsx components
