+++ package.json
@@ -17,7 +17,6 @@
     "express": "^4.18.2",
     "file-saver": "^2.0.5",
     "install": "^0.13.0",
-    "nodemon": "^2.0.20",
     "npm": "^9.1.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -34,6 +33,7 @@
     "babel-loader": "^9.1.0",
     "css-loader": "^6.7.2",
     "html-webpack-plugin": "^5.5.0",
+    "nodemon": "^2.0.20",
     "sass": "^1.56.1",
     "sass-loader": "^13.2.0",
     "style-loader": "^3.3.1",
