--- src/client/components/App.jsx
@@ -1,24 +1,0 @@
-import React from "react";
-import Header from "./Header.jsx";
-import Toolbar from "./Toolbar.jsx";
-import FetchWindow from "./FetchWindow.jsx";
-import EditorContainer from "./EditorContainer.jsx";
-import styles from '../styles/App.scss';
-
-class App extends React.Component {
-  render() {
-    return (
-      <div>
-        {/* <h1>App Component</h1> */}
-        <div id="header-components" style={styles}>
-          <Header />
-          <Toolbar />
-        </div>
-        <EditorContainer />
-        <FetchWindow />
-      </div>
-    );
-  }
-}
-
-export default App;
