+++ src/client/components/Header.jsx
@@ -7,7 +7,7 @@ class Header extends React.Component {
     return (
       <div id="react-header">
         <p>
-          <span id="filename">README.md </span>
+          <span id="filename">README.md</span>
           {/* <span id="src-repo">github.com/user/repo/blob/main/readme.md </span> */}
           {/* <span id="src-repo" hidden>github.com/{this.props.user}/{this.props.repo}/blob/{this.props.branch}/{this.props.filename}</span> */}
           <span id="src-repo">{this.props.url}</span>
