+++ src/client/containers/EditorContainer.jsx
@@ -1,6 +1,6 @@
 import React from "react";
-import Editor from "./Editor.jsx";
-import Preview from "./Preview.jsx";
+import Editor from "../components/Editor.jsx";
+import Preview from "../components/Preview.jsx";
 import '../styles/EditorContainer.scss';
 import { connect } from "react-redux";
 
@@ -8,7 +8,6 @@ class EditorContainer extends React.Component {
   render() {
     return (
       <div id="react-editor-container">
-        {/* <h1>Editor Container</h1> */}
         <Editor editorText={this.props.editorText} />
         <div id="arrow">&rarr;</div>
         <Preview previewText={this.props.previewText} />
