+++ src/client/main.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { render } from "react-dom";
 import { Provider } from "react-redux";
-import App from "./components/App.jsx";
+import App from "./App.jsx";
 import store from "./store.js";
 
 render(
